requirements.txt
python-telegram-bot==20.0a0
PyYAML==6.0
schedule
requests
pandas
tabulate
python-binance